requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
readchar==4.0.5
requests==2.31.0
rich==13.7.0
urllib3==2.0.6
python-dotenv==1.0.1
piexif==1.1.3
